Export agent metadata
